# application/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: JobsHunter API
  version: "1.0.0"
  description: |
    HTTP API for user authentication, job search, profile management, CV upload, and admin email tools.
servers:
  - url: https://ronin-archesporial-lonnie.ngrok-free.dev
    description: Prod
  - url: http://localhost:8081
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [oldPassword, newPassword]
    EngineType:
      type: string
      enum: [GPT, GEMINI, SERP]
...
    UserInfoResponse:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        notifyWhatsapp:
          type: boolean
        notifyEmail:
          type: boolean
        emailVerified:
          type: boolean
        verificationToken:
          type: string
          nullable: true
        gptFileId:
          type: string
          nullable: true
        geminiFileId:
          type: string
          nullable: true
        lastJobs:
          type: string
          nullable: true
          description: ISO-8601 date-time as string.
        timeInterval:
          type: integer
          format: int32
          nullable: true
        prompts:
          type: array
          nullable: true
          items:
            type: string
        createdAt:
          type: string
          nullable: true
          description: ISO-8601 date-time as string.
        roles:
          type: array
          items:
            type: string
      required:
        - username
        - email
        - phoneNumber
        - notifyWhatsapp
        - notifyEmail
        - emailVerified
        - roles
    UserJobResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
        createdAt:
          type: string
          description: ISO-8601 date-time as string.
      required: [url, createdAt]
...
    SerpPrompt:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
          description: Prompt ID for updates (null for new prompts)
        q:
          type: string
          maxLength: 2000
          description: SERP query.
        google_domain:
          type: string
          maxLength: 255
          nullable: true
          description: Google domain (e.g., google.ro). See https://serpapi.com/google-domains
        date_posted:
          type: string
          maxLength: 128
          nullable: true
          description: Date filter (today, last_3_days, last_7_days, last_14_days, last_30_days).
        gl:
          type: string
          maxLength: 2
          nullable: true
          description: Country code. See https://serpapi.com/google-countries
        location:
          type: string
          nullable: true
        workType:
          type: string
          enum: [REMOTE, ONSITE, HYBRID]
          nullable: true
        radius:
          type: integer
          format: int32
          nullable: true
          description: Radius in km.
        hl:
          type: string
          maxLength: 2
          nullable: true
          description: Language code. See https://serpapi.com/google-languages
      required: [q]
    AiPrompt:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
          description: Prompt ID for updates (null for new prompts)
        prompt:
          type: string
          maxLength: 3000
        engine:
          $ref: '#/components/schemas/EngineType'
        model:
          type: string
          maxLength: 255
          description: Model name (e.g., gpt-4o-mini, gemini-2.5-flash)
      required: [prompt, engine, model]
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
        phone_number:
          type: string
          maxLength: 30
        time_interval:
          type: integer
          format: int32
          minimum: 1
        notify_whatsapp:
          type: boolean
        notify_email:
          type: boolean
        serp_prompts:
          type: array
          items:
            $ref: '#/components/schemas/SerpPrompt'
        ai_prompts:
          type: array
          items:
            $ref: '#/components/schemas/AiPrompt'
      required:
        - username
        - phone_number
        - time_interval
...
paths:
  /api/user/delete:
    delete:
      summary: Delete the authenticated user's account
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/approve:
    patch:
      summary: Approve user by username (admin only)
      tags: [User]
      description: Requires ROLE_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
        - in: query
          name: rejectReason
          required: false
          schema:
            type: string
            maxLength: 255
          description: Optional rejection reason. If provided, sends rejection email instead of approval.
      responses:
        '200':
          description: User approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or already approved user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not test role).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/user/me:
    get:
      summary: Get current authenticated user profile
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/all:
    get:
      summary: List all users (admin only)
      tags: [User]
      description: Requires ROLE_ADMIN.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not test role).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/jobs:
    get:
      summary: Get jobs found for a user (admin only)
      tags: [User]
      description: Requires ROLE_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User jobs list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserJobResponse'
        '400':
          description: Missing or invalid username.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not test role).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/update:
    put:
      summary: Update user profile, notification flags, and prompts
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/password:
    patch:
      summary: Change password for authenticated user
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed; verification token sent by email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...